requirements.txt
streamlit
kagglehub
pillow
numpy
opencv-python
